% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/zoneparser.r
\name{parse_zone}
\alias{parse_zone}
\title{Parse a Domain Name System (DNS) zone file}
\usage{
parse_zone(path)
}
\arguments{
\item{path}{path to DNS zone file to parse}
}
\value{
\code{list} with DNS zone parsed
}
\description{
Parse a Domain Name System (DNS) zone file
}
\examples{
parse_zone(system.file("zones/20151001-wtf-zone-data.txt", package="zoneparser"))
}

