% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/zoneparser-package.r
\docType{package}
\name{zoneparser}
\alias{zoneparser}
\alias{zoneparser-package}
\title{A package to parse Domain Name System (DNS) zone files}
\description{
A package to parse Domain Name System (DNS) zone files
}
\author{
Bob Rudis (@hrbrmstr)
}

